fix: rename `PaginationHeader` to `PaginationAmount`
